Change database and name to H2 for class Student,User,Book. instruct hibernate  to validate data

+++ src/main/java/home/pratice/design/pattern/dao/BookDAO.java
@@ -3,6 +3,6 @@ package home.pratice.design.pattern.dao;
 import home.pratice.domain.Book;
 
 public interface BookDAO {
-    Book getBook(int bookId);
+    Book getBook(Long bookId);
     void saveBook(int bookId ,String bookName);
 }
